# add ignoreUnknownKeys option to record runtype