--- a/src/index.ts
+++ b/src/index.ts
@@ -858,17 +858,13 @@ export function numberIndex<T>(t: Runtype<T>): Runtype<{ [key: number]: T }> {
 
 /**
  * An object with defined keys and values.
- *
- * Options:
- *
- *   ignoreUnknownKeys .. will not complain about invalid keys in record
  */
-export function record<T extends object>(
+function internalRecord<T extends object>(
   typemap: { [K in keyof T]: Runtype<T[K]> },
-  options?: { ignoreUnknownKeys?: boolean },
+  sloppy: boolean,
 ): Runtype<T> {
   const isPure = Object.values(typemap).every((t: any) => isPureRuntype(t))
-  const { ignoreUnknownKeys } = options ?? {}
+  const copyObject = sloppy || !isPure
 
   const rt: Runtype<T> = internalRuntype((v, failOrThrow) => {
     const o: any = (objectRuntype as InternalRuntype)(v, failOrThrow)
@@ -880,7 +876,7 @@ export function record<T extends object>(
     // TODO: optimize allocations: only create a copy if any of the key
     // runtypes return a different object - otherwise return value as is
 
-    const res = isPure ? (o as T) : ({} as T)
+    const res = copyObject ? ({} as T) : (o as T)
 
     for (const k in typemap) {
       // nested types should always fail with explicit `Fail` so we can add additional data
@@ -890,12 +886,12 @@ export function record<T extends object>(
         return propagateFail(failOrThrow, value, v, k)
       }
 
-      if (!isPure) {
+      if (copyObject) {
         res[k] = value
       }
     }
 
-    if (!ignoreUnknownKeys) {
+    if (!sloppy) {
       const unknownKeys = Object.keys(o).filter(
         (k) => !Object.prototype.hasOwnProperty.call(typemap, k),
       )
@@ -925,11 +921,17 @@ export function record<T extends object>(
   return rt
 }
 
-// TODO:
-//   export function sloppyRecord<T extends object>(
-//     typemap: { [K in keyof T]: Runtype<T[K]> },
-//   ): Runtype<T>
-// same as `record` but without the 'unknownKeys' validation
+export function record<T extends object>(
+  typemap: { [K in keyof T]: Runtype<T[K]> },
+): Runtype<T> {
+  return internalRecord(typemap, false)
+}
+
+export function sloppyRecord<T extends object>(
+  typemap: { [K in keyof T]: Runtype<T[K]> },
+): Runtype<T> {
+  return internalRecord(typemap, true)
+}
 
 /**
  * A runtype based on a type guard