Later I noticed this comment
https://github.com/hoeck/simple-runtypes/blob/7fe99ed25727a20637fcf4b5a1258849ab0847d8/src/index.ts#L920-L924

Do you prefer the record runtype with options (as in my PR) or do you prefer a separate runtype sloppyRecord with duplicated logic?