@hoeck 2020-08-08
ATM, I would prefer `record.sloppy` :dagger: 

One reason is that the standard prettier formatting for a function that has two literal object parameters consumes way more space compared to that of a function that just takes object literal parameter.

Also I don't expect `record` to get any additional option parameters in the future.

Regardin the implementation, to keep the 'sloppy' record *safe* it should always build a new object with only the keys in it which are defined in the record runtype. This way, sloppy records keep you safe from attackers passing random keys through the runtype validation layer. What do you think?