@pabra 2020-08-09
Results of `typescript-runtime-type-benchmarks`

before:
```
++ node -v
+ export NODE_VERSION=v12.18.0
+ NODE_VERSION=v12.18.0
+ npm start

> typescript-runtime-type-benchmarks@1.0.0 start /home/pepp/git/typescript-runtime-type-benchmarks
> ts-node index.ts

Running "data-type" suite...
[...]
  simple-runtypes:
    1 334 501 ops/s, ±2.15%    | 95.75% slower
[...]
Running "data-type-sans-outliers" suite...
[...]
  simple-runtypes:
    1 352 364 ops/s, ±0.98%   | 33.89% slower
[...]
```

after:
```
++ node -v
+ export NODE_VERSION=v12.18.0
+ NODE_VERSION=v12.18.0
+ npm start

> typescript-runtime-type-benchmarks@1.0.0 start /home/pepp/git/typescript-runtime-type-benchmarks
> ts-node index.ts

Running "data-type" suite...
[...]
  simple-runtypes:
    1 356 510 ops/s, ±0.47%    | 95.61% slower
[...]
Running "data-type-sans-outliers" suite...
[...]
  simple-runtypes:
    1 355 933 ops/s, ±2.70%   | 31.1% slower
[...]
```